extract global functions
